+++ meerkat-contracts/src/error/mod.rs
@@ -284,6 +284,7 @@ impl From<meerkat_core::SessionError> for WireError {
             | meerkat_core::SessionError::CompactionDisabled
             | meerkat_core::SessionError::Unsupported(_) => ErrorCode::CapabilityUnavailable,
             meerkat_core::SessionError::DurableTailHeldForRecovery { .. }
+            | meerkat_core::SessionError::DurableTailRecoveryRefused { .. }
             | meerkat_core::SessionError::DurableEvidenceQuarantined { .. } => {
                 ErrorCode::SessionNotRunning
             }